Add async exaple & split async integration tests
* Provide simple async mpsc exmaple to help users know its APIs.
* Split the async tests to be more clear on what integration tests are
testing.
